Add color npm package, decompose EventChat direct messages components, clean up styling a bit more to reflect latest design direction, fix overrides for Messaging chat theme
