+++ src/styles/theme.js
@@ -1,3 +1,5 @@
+import color from 'color';
+
 // Corresponds to bootstrap's values for styles like p-1, mr-2, px-3, etc.
 // Notice the values are not linear.
 const baseUnitMap = Object.freeze([
@@ -50,19 +52,11 @@ const colors = Object.freeze({
  * the color mapping exists, doesn't mean it always makes sense to use it.
  */
 export const theme = Object.freeze({
+  // Globals
   brand: colors.cyan,
-  body: {
-    background: colors.gray[50],
-  },
-  card: {
-    background: colors.white,
-    color: colors.gray[500],
-  },
   font: {
     ...colors.gray, // Hypothetically, this could be reversed for dark mode
   },
-  link: colors.cyan,
-  liveEvent: colors.red,
   shadow: {
     small: '0 0.125rem 0.25rem rgba(0, 0, 0, 0.075)',
     medium: '0 0.5rem 1rem rgba(0, 0, 0, 0.15)',
@@ -76,4 +70,22 @@ export const theme = Object.freeze({
     circle: '50%',
     pill: '50rem',
   },
+
+  // UI Elements
+  body: {
+    background: colors.gray[50],
+  },
+  card: {
+    background: colors.white,
+    color: colors.gray[500],
+  },
+  chat: {
+    dmsHeader: color(colors.white).alpha(0.9).rgb(),
+    message: {
+      name: colors.gray[700],
+      text: colors.gray[600],
+    },
+  },
+  link: colors.cyan,
+  liveEvent: colors.red,
 });
